% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_assign_clones_to_cells.R
\name{assign_clone_by_edit_distance}
\alias{assign_clone_by_edit_distance}
\title{Assign clones by edit distance.}
\usage{
assign_clone_by_edit_distance(
  cells_to_find,
  cell_clone_reads_dt,
  cell_bcode_col,
  barcode_edit_dist_col,
  clone_bcode_col
)
}
\arguments{
\item{cells_to_find}{A vector of multiclone cells to do the assignment.}

\item{cell_clone_reads_dt}{A data.table object representing the reads. Each
row includes information about a cell, UMI, and clone barcode. For data
produced by NextClone, use `fread` from the data.table package and pass the
resulting object to this parameter.}

\item{cell_bcode_col}{Name of the column in `cell_clone_reads_dt` that
indicates the cell barcode for each read.}

\item{barcode_edit_dist_col}{Name of the column in `cell_clone_reads_dt` that
indicates the edit distance for the clone barcode of each read.}

\item{clone_bcode_col}{Name of the column in `cell_clone_reads_dt` that
specifies the clone barcode for each read.}
}
\value{
A data.table denoting the clones for multiclone cells.
}
\description{
Internal function to assign clones to multiclone cells based on the edit
distance of the clone barcode reads
}
\keyword{internal}
