% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_ordered_abundance_plot.R
\name{assign_numeric_barcode_id}
\alias{assign_numeric_barcode_id}
\title{Assign numeric barcode id}
\usage{
assign_numeric_barcode_id(
  count_data,
  count_column,
  elbow_grouping_col = NA,
  plot_grouping_col = NA
)
}
\arguments{
\item{count_data}{A data.table where each row corresponds to a clone barcode.}

\item{count_column}{A character indicating the column storing the count of the clone barcodes.}

\item{elbow_grouping_col}{A character string, default NA, indicating the column which values separates each elbow.}

\item{plot_grouping_col}{A character string, default NA, indicating the column which values separates each plot.}
}
\value{
A data.table with numerical ID for each clone barcode.
}
\description{
For each clone barcode assign a numeric id.
IDs are ordered based on the number of cells for each clone (clone with the largest count has small ID).
}
\keyword{internal}
