% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_elbow_plot.R
\name{assign_numeric_barcode_id_multiple_elbows}
\alias{assign_numeric_barcode_id_multiple_elbows}
\title{Assign numeric barcode id for all elbows}
\usage{
assign_numeric_barcode_id_multiple_elbows(
  all_elbow_data,
  elbow_grouping_col,
  count_column
)
}
\arguments{
\item{all_elbow_data}{A data.table storing the data for all elbows.}

\item{elbow_grouping_col}{A column in the data.table that separate each elbow.}

\item{count_column}{A character indicating the column storing the count of the clone barcodes.}
}
\value{
A new data.table
}
\description{
If a plot has multiple elbows, assign the numerical barcode id such that
for the first elbow, the barcode id starts from 0,
then the subsequent elbows, the barcode id starts from the end of the previous elbow.
}
