% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_clones.R
\name{convert_count_to_proportion}
\alias{convert_count_to_proportion}
\title{Convert count to proportion}
\usage{
convert_count_to_proportion(count_data, count_column, grouping_col = NA)
}
\arguments{
\item{count_data}{A data.table containing clone barcode count data.}

\item{count_column}{A character string specifying the column storing clone barcode counts.}

\item{grouping_col}{A character string, default NA. If provided, the projection
is repeated for each group defined by this column.}
}
\value{
A modified data.table with a new column representing the proportion.
}
\description{
For each clone barcode, convert the absolute count to proportion.
If `grouping_col` is specified, the proportion will be calculated with respect
to the total number of cells in each group.
}
\examples{
toy_clone_counts <- data.table(
sample_name = c(rep("test1", 2), rep("test2", 3)),
read_count = c(1, 5, 10, 15, 20)
)

convert_count_to_proportion(
    count_data = toy_clone_counts,
    grouping_col = "sample_name",
    count_column = "read_count"
)
}
