% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_clones.R
\name{count_retained_clones}
\alias{count_retained_clones}
\title{Calculate clone barcode with at least x number of cells}
\usage{
count_retained_clones(count_data, thresholds, count_column, grouping_col = NA)
}
\arguments{
\item{count_data}{A data.table where each row corresponds to a clone barcode.}

\item{thresholds}{A numeric vector listing the minimum cell counts required for a clone to be included in the count.
The calculation will be applied for each value in the vector.}

\item{count_column}{A character indicating the column storing the count of the clone barcodes.}

\item{grouping_col}{A character string, default NA. If provided, the function repeats the calculation for each subsets of `count_data` defined by this column, effectively a group-by operation.}
}
\value{
A data.table containing the counts of clones that meet or exceed the specified threshold.
}
\description{
This function calculates the number of clones with at least a specified number of cells.
Clones with fewer cells than the threshold are excluded from the count.
}
\examples{
toy_clone_counts <- data.table(
sample_name = c(rep("test1", 2), rep("test2", 3)),
clone_barcode = c("A", "B", "C", "D", "E"),
read_count = c(1, 5, 10, 15, 20)
)

count_retained_clones(
  count_data = toy_clone_counts,
  thresholds = c(2, 10),
  count_column = "read_count",
  grouping_col = "sample_name"
)

}
