% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_ordered_abundance_plot.R
\name{draw_ordered_abundance_plot}
\alias{draw_ordered_abundance_plot}
\title{Draw ordered abundance plot}
\usage{
draw_ordered_abundance_plot(
  count_data,
  y_axis_column,
  colour_column = NA,
  facet_column = NA,
  facet_row = NA
)
}
\arguments{
\item{count_data}{A data.table where each row corresponds to a clone barcode.}

\item{y_axis_column}{The column which data to display in y-axis.
Note, the numeric barcode ID will also be ordered based on this.}

\item{colour_column}{A character string, default NA. If provided, indicates the column to colour the plot by.}

\item{facet_column}{A character string, default NA. If provided, indicates the column which value separates the plots into column.}

\item{facet_row}{A character string, default NA. If provided, indicates the column which value separates the plots into row}
}
\value{
A ggplot object
}
\description{
The ordered abundance plot shows the proportion of barcodes' frequency.
}
\examples{
library(data.table)

toy_clone_counts <- data.table(
    treatment = c(rep("treat1", 4), rep("treat2", 4)),
    replicate = rep(c(rep("rep1", 2), rep("rep2", 2)), 2),
    read_count = c(20, 10, 15, 2, 3, 30, 15, 20)
)
draw_ordered_abundance_plot(
    count_data = toy_clone_counts,
    count_column = "read_count",
    facet_column = "replicate",
    facet_row = "treatment",
    y_axis_column = 'read_count'
)

}
