% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_draw_treemap.R
\name{draw_treemap}
\alias{draw_treemap}
\title{Draw treemap}
\usage{
draw_treemap(
  cell_by_clone_matrix,
  valid_cells_bcodes,
  cell_bcode_col = "CellBarcode"
)
}
\arguments{
\item{cell_by_clone_matrix}{A data.table where each row denotes a cell, a clone barcode,
and the number of reads assigned to the clone.}

\item{valid_cells_bcodes}{A character vector of cell barcodes for droplets
deemed real cells.
Can be the cell barcodes in cell by gene matrix generated by cellranger's
filtered count matrix.}

\item{cell_bcode_col}{The column in cell_by_clone_matrix denoting the cell's barcode.}
}
\value{
A ggplot object containing the treemap.
}
\description{
Draw treemap which shows the number and percentage of cells with 0, 1, 2, etc.
clone barcodes.
}
\examples{
library(data.table)

cell_by_clone_matrix <- data.table(
    CellBarcode = c("A", "B", "C", "C"),
    CloneBarcode = c("AA", "AA", "BB", "CC"),
    n_reads = as.integer(c(2, 2, 2, 1))
)
draw_treemap(
    cell_by_clone_matrix = cell_by_clone_matrix,
    valid_cells_bcodes = c("A", "B", "C")
)

}
