% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_assign_clones_to_cells.R
\name{find_and_rank_clones_in_multiclone_cells}
\alias{find_and_rank_clones_in_multiclone_cells}
\title{Count number of clones per cells.}
\usage{
find_and_rank_clones_in_multiclone_cells(
  clone_count_per_cell,
  cell_bcode_col,
  cell_by_clone_matrix
)
}
\arguments{
\item{clone_count_per_cell}{A data.table denoting the number of clones detected
per cell. Generated by `count_clones_per_cells` function.}

\item{cell_bcode_col}{Name of the column in `cell_by_clone_matrix` that
indicates the cell barcode.}

\item{cell_by_clone_matrix}{A cell-by-clone matrix generated by
`generate_cell_clone_barcode_matrix` function.}
}
\value{
A data.table denoting the ranking of clones for multiclone cells.
}
\description{
Internal function to find cells with multiple clones detected,
and rank the clones based on the proportion of reads detected for them.
}
\keyword{internal}
