% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_assign_clones_to_cells.R
\name{get_cells_clones_assignment}
\alias{get_cells_clones_assignment}
\title{Get clone assignments.}
\usage{
get_cells_clones_assignment(
  valid_cells_bcodes,
  cell_clone_reads_dt,
  cell_bcode_col = "CellBarcode",
  barcode_edit_dist_col = "BarcodeEditDist",
  clone_bcode_col = "CloneBarcode",
  umi_col = "UMI",
  most_dominant_threshold = 0.5,
  umi_clone_consensus_threshold = 0.7
)
}
\arguments{
\item{valid_cells_bcodes}{A vector of cells barcodes to get the clone
assignments for.}

\item{cell_clone_reads_dt}{A data.table object representing the reads. Each
row includes information about a cell, UMI, and clone barcode. For data
produced by NextClone, use `fread` from the data.table package and pass the
resulting object to this parameter.}

\item{cell_bcode_col}{Name of the column in `cell_clone_reads_dt` that
indicates the cell barcode for each read.}

\item{barcode_edit_dist_col}{Name of the column in `cell_clone_reads_dt` that
indicates the edit distance for the clone barcode of each read.}

\item{clone_bcode_col}{Name of the column in `cell_clone_reads_dt` that
specifies the clone barcode for each read.}

\item{umi_col}{Name of the column in `cell_clone_reads_dt` that specifies the
UMI barcode for each read.}

\item{most_dominant_threshold}{A numeric value between 0 and 1 specifying the
proportion threshold of reads that must be associated with a clone barcode to
assign it to a cell in cases where multiple clone barcodes are detected.}

\item{umi_clone_consensus_threshold}{A numeric value between 0 and 1 specifying
the proportion threshold of reads for collapsing UMIs when computing the
cell-by-clone matrix.
See details for more information.}
}
\value{
A data.table of cell and clone assignment.
}
\description{
Internal function which apply the stratified protocol described in
`assign_and_embed_clones` function.
}
\keyword{internal}
