% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_clones.R
\name{get_top_barcodes_and_cum_sum}
\alias{get_top_barcodes_and_cum_sum}
\title{Get top barcodes and cumulative sum of the counts.}
\usage{
get_top_barcodes_and_cum_sum(
  count_data,
  top_threshold,
  count_column,
  grouping_col = NA
)
}
\arguments{
\item{count_data}{A data.table where each row corresponds to a clone barcode.}

\item{top_threshold}{A numeric value indicating how many of the top clone barcodes to retain.}

\item{count_column}{A character string indicating the column storing the count of the clone barcodes.}

\item{grouping_col}{A character string, default NA. If provided, the function repeats the calculation for each subset of `count_data` defined by this column, effectively performing a group-by operation.}
}
\value{
A filtered data.table with only the top clone barcodes.
}
\description{
Filter the count data to retain only the top x barcodes, where x is defined by top_threshold.
}
\examples{
library(data.table)

toy_clone_counts <- data.table(
sample_name = c(rep("test1", 3), rep("test2", 3)),
read_count = c(1, 5, 20, 10, 15, 12),
clone_barcodes = c("ACGT", "CATG", "CATG", "ACGT", "ATGC", "TCGT")
)

get_top_barcodes_and_cum_sum(
  count_data = toy_clone_counts,
  top_threshold = 2,
  count_column = "read_count",
  grouping_col = "sample_name"
)
}
