% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_mds_plot.R
\name{plot_mds_as_ggplot}
\alias{plot_mds_as_ggplot}
\title{Draw MDS plot using ggplot.}
\usage{
plot_mds_as_ggplot(dge_list, lib_sizes)
}
\arguments{
\item{dge_list}{A list of DGEList objects, 1 per library.}

\item{lib_sizes}{A character vector of library sizes.
1 indexed element corresponds to 1 indexed element in `dge_list`.}
}
\description{
The default plotMDS from edgeR uses R's `plot` function which is hard
to manipulate. This function will draw MDS plot as ggplot and arrange
it into a grid broken down by library sizes.
}
