% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecting_clones.R
\name{projecting_clones}
\alias{projecting_clones}
\title{Calculate clone projection}
\usage{
projecting_clones(
  count_data,
  count_column,
  grouping_col = NA,
  project_amnt = 10000
)
}
\arguments{
\item{count_data}{A data.table containing clone barcode count data.}

\item{count_column}{A character string specifying the column storing clone barcode counts.}

\item{grouping_col}{A character string, default NA. If provided, the projection
is repeated for each group defined by this column.}

\item{project_amnt}{A numeric value indicating the number of cells to project to.}
}
\value{
A modified data.table with projected cell counts.
}
\description{
Predicts the number of cells tagged with each clone barcode for a given projection amount.
This function is useful for planning single-cell experiments.
It allows you to estimate the number of cells associated with specific clone barcodes
and anticipate the abundance of clone barcodes (i.e., how many clone barcodes
with a substantial number of cells).
}
\details{
How this projection is done: For each clone barcode, we compute their proportion,
and then multiply the proportion by the number of cells to project to
(specified by the `project_amnt` parameter).
If `grouping_col` is specified, the proportion will be calculated with respect
to the total number of cells in each group.
}
\examples{
toy_clone_counts <- data.table(
sample_name = c(rep("test1", 3), rep("test2", 3)),
read_count = c(1, 5, 20, 10, 15, 12),
clone_barcodes = c("ACGT", "CATG", "CATG", "ACGT", "ATGC", "TCGT")
)

projecting_clones(
    count_data = toy_clone_counts,
    project_amnt = 100,
    count_column = "read_count",
    grouping_col = "sample_name"
)
}
